+++ changelog.md
@@ -1,2 +1,4 @@
 v0.1.0 initial version
 v0.1.1 create player works
+v0.1.2 creature, inventory & item classes works.
+       add pytest tests